refactor: CompositeEvaluatorFactory refactor
